moved conversion functions to geometry_util
