+++ test_autoarray/geometry/test_geometry_util.py
@@ -1185,3 +1185,172 @@ def test__grid_pixel_centres_2d_slim_from():
     )
 
     assert (grid_pixels == np.array([[0, 0], [0, 1], [1, 0], [1, 1]])).all()
+
+
+def test__grid_pixel_indexes_2d_slim_from():
+
+    # coordinates in centres of pixels
+
+    grid_scaled = np.array([[1.0, -2.0], [1.0, 2.0], [-1.0, -2.0], [-1.0, 2.0]])
+
+    grid_pixels = aa.util.geometry.grid_pixel_indexes_2d_slim_from(
+        grid_scaled_2d_slim=grid_scaled,
+        shape_native=(2, 2),
+        pixel_scales=(2.0, 4.0),
+    )
+
+    assert (grid_pixels == np.array([0, 1, 2, 3])).all()
+
+    # coordinates top-left of pixels
+
+    grid_scaled = np.array([[1.99, -3.99], [1.99, 0.01], [-0.01, -3.99], [-0.01, 0.01]])
+
+    grid_pixels = aa.util.geometry.grid_pixel_indexes_2d_slim_from(
+        grid_scaled_2d_slim=grid_scaled,
+        shape_native=(2, 2),
+        pixel_scales=(2.0, 4.0),
+    )
+
+    assert (grid_pixels == np.array([0, 1, 2, 3])).all()
+
+    # coordinates bottom-right of pixels
+
+    grid_scaled = np.array([[0.01, -0.01], [0.01, 3.99], [-1.99, -0.01], [-1.99, 3.99]])
+
+    grid_pixels = aa.util.geometry.grid_pixel_indexes_2d_slim_from(
+        grid_scaled_2d_slim=grid_scaled,
+        shape_native=(2, 2),
+        pixel_scales=(2.0, 4.0),
+    )
+
+    assert (grid_pixels == np.array([0, 1, 2, 3])).all()
+
+    # non-zero origin
+
+    # +1.0 for all entries for a origin of (1.0, 1.0)
+    grid_scaled = np.array([[2.0, -1.0], [2.0, 3.0], [0.0, -1.0], [0.0, 3.0]])
+
+    grid_pixels = aa.util.geometry.grid_pixel_indexes_2d_slim_from(
+        grid_scaled_2d_slim=grid_scaled,
+        shape_native=(2, 2),
+        pixel_scales=(2.0, 4.0),
+        origin=(1.0, 1.0),
+    )
+
+    assert (grid_pixels == np.array([0, 1, 2, 3])).all()
+
+
+def test__grid_scaled_2d_slim_from():
+
+    # coordinates in centres of pixels
+
+    grid_pixels = np.array([[0.5, 0.5], [0.5, 1.5], [1.5, 0.5], [1.5, 1.5]])
+
+    grid_scaled = aa.util.geometry.grid_scaled_2d_slim_from(
+        grid_pixels_2d_slim=grid_pixels,
+        shape_native=(2, 2),
+        pixel_scales=(2.0, 4.0),
+    )
+
+    assert (
+        grid_scaled == np.array([[1.0, -2.0], [1.0, 2.0], [-1.0, -2.0], [-1.0, 2.0]])
+    ).all()
+
+    # coordinates top-left of pixels
+
+    grid_pixels = np.array([[0, 0], [0, 1], [1, 0], [1, 1]])
+
+    grid_scaled = aa.util.geometry.grid_scaled_2d_slim_from(
+        grid_pixels_2d_slim=grid_pixels,
+        shape_native=(2, 2),
+        pixel_scales=(2.0, 4.0),
+    )
+
+    assert (
+        grid_scaled == np.array([[2.0, -4], [2.0, 0.0], [0.0, -4], [0.0, 0.0]])
+    ).all()
+
+    # coordinates bottom-right of pixels
+
+    grid_pixels = np.array(
+        [[0, 0], [0, 1], [0, 2], [1, 0], [1, 1], [1, 2], [2, 0], [2, 1], [2, 2]]
+    )
+
+    grid_pixels = np.array([[1, 1], [1, 2], [2, 1], [2, 2]])
+
+    grid_scaled = aa.util.geometry.grid_scaled_2d_slim_from(
+        grid_pixels_2d_slim=grid_pixels,
+        shape_native=(2, 2),
+        pixel_scales=(2.0, 4.0),
+    )
+
+    assert (
+        grid_scaled == np.array([[0.0, 0.0], [0.0, 4.0], [-2.0, 0.0], [-2.0, 4.0]])
+    ).all()
+
+    # non-zero origin
+
+    grid_pixels = np.array([[0.5, 0.5], [0.5, 1.5], [1.5, 0.5], [1.5, 1.5]])
+
+    grid_scaled = aa.util.geometry.grid_scaled_2d_slim_from(
+        grid_pixels_2d_slim=grid_pixels,
+        shape_native=(2, 2),
+        pixel_scales=(2.0, 4.0),
+        origin=(-1.0, -1.0),
+    )
+
+    # -1.0 from all entries for a origin of (-1.0, -1.0)
+    assert (
+        grid_scaled == np.array([[0.0, -3.0], [0.0, 1.0], [-2.0, -3.0], [-2.0, 1.0]])
+    ).all()
+
+
+def test__grid_pixel_centres_2d_from():
+
+    # coordinates in centres of pixels
+
+    grid_scaled = np.array([[[1.0, -2.0], [1.0, 2.0]], [[-1.0, -2.0], [-1.0, 2.0]]])
+
+    grid_pixels = aa.util.geometry.grid_pixel_centres_2d_from(
+        grid_scaled_2d=grid_scaled, shape_native=(2, 2), pixel_scales=(2.0, 4.0)
+    )
+
+    assert (grid_pixels == np.array([[[0, 0], [0, 1]], [[1, 0], [1, 1]]])).all()
+
+    # coordinates top-left of pixels
+
+    grid_scaled = np.array(
+        [[[1.99, -3.99], [1.99, 0.01]], [[-0.01, -3.99], [-0.01, 0.01]]]
+    )
+
+    grid_pixels = aa.util.geometry.grid_pixel_centres_2d_from(
+        grid_scaled_2d=grid_scaled, shape_native=(2, 2), pixel_scales=(2.0, 4.0)
+    )
+
+    assert (grid_pixels == np.array([[[0, 0], [0, 1]], [[1, 0], [1, 1]]])).all()
+
+    # coordinates bottom-right of pixels
+
+    grid_scaled = np.array(
+        [[[0.01, -0.01], [0.01, 3.99]], [[-1.99, -0.01], [-1.99, 3.99]]]
+    )
+
+    grid_pixels = aa.util.geometry.grid_pixel_centres_2d_from(
+        grid_scaled_2d=grid_scaled, shape_native=(2, 2), pixel_scales=(2.0, 4.0)
+    )
+
+    assert (grid_pixels == np.array([[[0, 0], [0, 1]], [[1, 0], [1, 1]]])).all()
+
+    # non-zero origin
+
+    # +1.0 for all entries for a origin of (1.0, 1.0)
+    grid_scaled = np.array([[[2.0, -1.0], [2.0, 3.0]], [[0.0, -1.0], [0.0, 3.0]]])
+
+    grid_pixels = aa.util.geometry.grid_pixel_centres_2d_from(
+        grid_scaled_2d=grid_scaled,
+        shape_native=(2, 2),
+        pixel_scales=(2.0, 4.0),
+        origin=(1.0, 1.0),
+    )
+
+    assert (grid_pixels == np.array([[[0, 0], [0, 1]], [[1, 0], [1, 1]]])).all()
